fuse separated into separate header
